refactor: group map interaction handlers (#8336)
## Summary
- move the existing map interaction handler sources into the
`map/handlers` directory alongside the custom line handler
- update include directives and build configuration entries to reference
the new locations

## Testing
- not run (not requested)

------
https://chatgpt.com/codex/tasks/task_e_68f188608150832ab0925e8175eccedf
