+++ src/map/handlers/LabelInteractionHandler.h
