+++ src/map/handlers/PanInteractionHandler.h
