+++ src/map/handlers/SelectionRectangleHandler.h
