+++ src/mudlet.pro
@@ -664,9 +664,9 @@ SOURCES += \
     ShortcutsManager.cpp \
     SingleLineTextEdit.cpp \
     T2DMap.cpp \
-    LabelInteractionHandler.cpp \
-    PanInteractionHandler.cpp \
-    SelectionRectangleHandler.cpp \
+    map/handlers/LabelInteractionHandler.cpp \
+    map/handlers/PanInteractionHandler.cpp \
+    map/handlers/SelectionRectangleHandler.cpp \
     TAccessibleTextEdit.cpp \
     TAction.cpp \
     TAlias.cpp \
@@ -802,9 +802,9 @@ HEADERS += \
     ShortcutsManager.h \
     SingleLineTextEdit.h \
     T2DMap.h \
-    LabelInteractionHandler.h \
-    PanInteractionHandler.h \
-    SelectionRectangleHandler.h \
+    map/handlers/LabelInteractionHandler.h \
+    map/handlers/PanInteractionHandler.h \
+    map/handlers/SelectionRectangleHandler.h \
     TAccessibleConsole.h \
     TAccessibleTextEdit.h \
     TAction.h \
